[#32] Failure when adding a region and segment
The GUI now forces region and segment names to be uppercase in the
model and in the display as only uppercase region and segment names
are valid from the user point of view.

Also, the webservices for save, verify, and delete are modified to
uppercase the region and segment name and for regions the DYNAMIC_SEGMENT
value.