[#29] Remove global usage
Add more checks to webserver code in places where globals are used.
This removes debugging and logging support when the application
is running. Setting the NOGBL flag to false will enable the debugging
and logging again. In the future the web server may log to stdout:
[https://github.com/shabiel/M-Web-Server/issues/23]

Not all places that use globals have NOGBL checks, only the places
where it is used in the GDE GUI application.

Delete GDUSEDEFS error from error array as it isn't a real errror
that should prevent GDE from running. This allows running the GUI
against a system that only has ydb_gbldir set with no gde file created