[#48] Journal option not saved
Journal Enabled isn't saved correctly due to sending the string "true"
or "false" to the web service which the code isn't designed to
understand - thus made no changes to the region/segment.

This actually affected all true/false checkboxes in region/segment
add/edit. Change the verify code to translate true/false to 1/0
which allows the changes to be saved correctly.