[#59] Convert save to also delete items
GDE needs to have deletes and additions within the same transaction
to properly handle certain use cases. This is accomplished by adding
a "deletedItems" object into the `save` call that follows the same
format as the `delete` call. Code is re-used from the `delete` call
to accomplish the task in the `save` call.

Add test to for `save` call to simulate adding and deleting a name
within a single transaction to verify functionality works as intended.